requirements.txt
numpy==1.19.5
pandas==1.1.5
pygame==2.1.2
PyOpenGL==3.1.6
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
